Revert "warnings be gone."
This reverts commit 5e2d54339a303b42be6e473401536dc5c99d5161.

+++ media/libstagefright/MediaCodec.cpp
@@ -125,7 +125,7 @@ status_t MediaCodec::init(const char *name, bool nameIsType, bool encoder) {
         if (codecIdx >= 0) {
             Vector<AString> types;
             if (mcl->getSupportedTypes(codecIdx, &types) == OK) {
-                for (size_t i = 0; i < types.size(); i++) {
+                for (int i = 0; i < types.size(); i++) {
                     if (types[i].startsWith("video/")) {
                         needDedicatedLooper = true;
                         break;
